+++ console-racing-mission/src/test/kotlin/domain/CarTest.kt
@@ -1,6 +1,7 @@
 package domain
 
 import com.model.Car
+import domain.service.FakeNumberGenerator
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
